7.3 start

+++ module7_2/src/main/java/entity/Currency.java
@@ -1,12 +1,14 @@
 package entity;
 
+import java.math.BigDecimal;
+
 public class Currency {
 
     private String abbreviation;
     private String name;
-    private double conversion_rate_to_EUR;
+    private BigDecimal conversion_rate_to_EUR;
 
-    public Currency(String abbreviation, String name , double d){
+    public Currency(String abbreviation, String name , BigDecimal d){
         this.abbreviation = abbreviation;
         this.name = name;
         this.conversion_rate_to_EUR = d;
@@ -29,11 +31,11 @@ public class Currency {
         this.name = name;
     }
 
-    public double getConversion_rate_to_EUR() {
+    public BigDecimal getConversion_rate_to_EUR() {
         return conversion_rate_to_EUR;
     }
 
-    public void setConversion_rate_to_EUR(double conversion_rate_to_EUR) {
+    public void setConversion_rate_to_EUR(BigDecimal conversion_rate_to_EUR) {
         this.conversion_rate_to_EUR = conversion_rate_to_EUR;
     }
